api refactoring
