+++ src/components/edit-article-form/edit-article-form.tsx
@@ -7,7 +7,8 @@ import { RouteComponentProps, withRouter } from 'react-router-dom'
 import TextArea from 'antd/es/input/TextArea'
 
 import { AppDispatch, DispatchType } from '../../services/store/store'
-import { RootState, editArticle, createArticle, Article } from '../../services/store/articles-slice'
+import { editArticle, createArticle } from '../../services/store/articles-slice'
+import { Article, RootState } from '../../services/api/articles-api-types'
 import FormTextField from '../form-components/form-text-field'
 import TagsBlock from '../form-components/tags-block'
 import { ITag, ITagable } from '../form-components/tags-block/tags-block'
